test(dialtone-docs): DLT-3268 refactor test suite — consolidate, remove hardcoded values, add CI workflow
