--- packages/dialtone-docs/tests/tests/searchability.test.js
@@ -1,50 +1,0 @@
-import { describe, test, expect, beforeAll } from 'vitest';
-import { findFiles, readFile } from '@helpers/fileReader.js';
-
-describe('Searchability — content uses actual names', () => {
-  let allContent;
-
-  beforeAll(async () => {
-    const files = await findFiles('src/content/**/*.md', { ignore: ['**/INDEX.md', '**/node_modules/**'] });
-    allContent = files.map(f => readFile(f)).join('\n');
-  });
-
-  test('documentation mentions actual package names', () => {
-    const content = allContent;
-    const packages = ['dialtone-vue', 'dialtone-css', 'dialtone-tokens', 'dialtone-icons'];
-
-    for (const pkg of packages) {
-      expect(content, `package name "${pkg}" not found in content`).toContain(pkg);
-    }
-  });
-
-  test('documentation mentions technical terms', () => {
-    const content = allContent;
-    const terms = ['pnpm', 'NX', 'monorepo', 'Vue'];
-
-    for (const term of terms) {
-      expect(content, `technical term "${term}" not found in content`).toContain(term);
-    }
-  });
-
-  test('no placeholder content in substantive files', async () => {
-    const files = await findFiles('src/content/**/*.md', { ignore: ['**/INDEX.md', '**/node_modules/**'] });
-    const badPatterns = [/\bTODO\b/, /\bFIXME\b/, /\bcoming soon\b/i, /\bTBD\b/];
-
-    for (const file of files) {
-      const content = readFile(file);
-      // Strip fenced code blocks and inline code before checking — standards
-      // docs legitimately mention "TODO" as examples of what NOT to do.
-      const stripped = content
-        .replace(/```[\s\S]*?```/g, '')
-        .replace(/`[^`]+`/g, '')
-        .replace(/"[^"]*"/g, '');
-      for (const pattern of badPatterns) {
-        expect(
-          pattern.test(stripped),
-          `${file} contains placeholder: ${pattern}`,
-        ).toBe(false);
-      }
-    }
-  });
-});
